Restructuring the database code.
Not separating the async code from the "lassic"code
